[lib][native][web] useIdentityWalletLogIn() hook
Summary:
hook that will be used on web and native for SIWE

Depends on D10960

Test Plan: used the hook on web to call the identity service. siwe verification failed as expected (used a random string for the siwe message and signature) but identity logs showed that the request received contained all the correct data. siwe will be tested thoroughly in the next diffs by actually using the hook in native and web siwe components

Reviewers: ashoat, inka, tomek

Differential Revision: https://phab.comm.dev/D10961