[lib] fix @-mentioning in a pending GENESIS sidebar crashes app
Summary: @-mentioning in a pending GENSIS sidebar would crash the app. This was because in `useMentionTypeaheadChatSuggestions` we were attempting to get the chat mention search index from the `communityThreadIDForGenesisThreads` with a pending thread info id. The reason this is an issue is that we build the `communityThreadIDForGenesisThreads`

Test Plan: Please see the demo video below

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10956