[lib] add prefix query support to context handler
Summary:
This implements prefix query messaging on the identity search context handler.
`sendPrefixQuery` add a promise to an array and then returns the promise. It then sends a websocket search query with an id.
When either a Success or Error search response type is received, the context handler will either resolve or reject promises using the id in the search response.

Test Plan: flow and console logged search results on web

Reviewers: varun, bartek, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11000