[lib] Introduce `ServerLegacyRoleInfo`
Summary:
This is the `Server`-equivalent of `ClientLegacyRoleInfo` which was renamed from `LegacyRoleInfo` in D10968.

> Why are we introducing a type without using it anywhere?

This type really should have been introduced as part of D10105. It will be immediately consumed by D10106 which will be sequenced after this diff, but didn't think it made sense to consolidate into one diff.

---

Depends on D10968

Test Plan: Will be tested implicitly via D10106 and by rest of stack.

Reviewers: ashoat, ginsu, tomek

Differential Revision: https://phab.comm.dev/D10969