[keyserver] Update thread-fetchers to fetch special_role
Summary:
We want to update code that fetches from `threads` or `roles` to appropriately also fetch `roles.special_role` so we can include it in the return object. For now, I will leave in the `r.special_role = ${specialRoles.DEFAULT_ROLE} AS is_default` since I'm intermediately leaving in the `isDefault` property in `RoleInfo`. Towards the end of the stack, I will remove this check alongside the property as mentioned in the Linear task.

Part of [[ https://linear.app/comm/issue/ENG-5994/introduce-specialrole-field-in-roleinfo | ENG-5993 ]]

This diff handles `thread.fetchers.js`

Depends on D10106

Test Plan: Ran the query before and after to confirm the results are identical, except with the new `special_role` field

Reviewers: ginsu, tomek, rohan

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10107