[lib] Remove support for `ServerThreadInfo` from `useKeyserverAdmin`
Summary:
In D10107 I encountered a scenario where `flow` wasn't able to type check `ServerLegacyRoleInfo` as part of a `ServerThreadInfo` being constructed.

To be safe, I decided to look at all instances of `ServerThreadInfo` and `fetchServerThreadInfos()` in the codebase to double check for similar issues.

I didn't find any, but noticed that `useKeyserverAdmin` accepted `ServerThreadInfo` which seems strange. Looks like it was added when `useKeyserverAdmin` was `getKeyserverAdmin`. Regardless, it's no longer necessary so I removed it.

---

Depends on D10107

Test Plan:
`flow` + seeing that the hook is only ever used by the `CommunityPill` react component which should never encounter a `ServerThreadInfo`.

Reviewers:
ashoat, ginsu, tomek