[web] make useSortedENSResolvedUsers more permissive
Summary:
In the next diff we are going to want to use `useSortedENSResolvedUsers` to sort + resolve the thread members infos just like we did with the add friend list modal. However, right now `useSortedENSResolvedUsers` is currently typed so thread members can't use this hook since they use `RelativeMemberInfo`.

This diff updates `useSortedENSResolvedUsers` so that it can be used in both places

Test Plan: `flow` + confirmed that users are sorted correctly in the add users list + thread members list

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10920