[web] cleanup change member role modal
Summary:
This diff cleans up the one off styles in the change member role modal and replaces it with styles of the redesigned modal. Additionally we fixed the issue where the height of the modal changes when the dropdown open/closes. To fix this, I made the height of the change member role modal match the height of the window. I did this so that the height of the modal was consistent with other full screen height modals like the pinned messages modal + message search modal for example. However, if we think this is too much white space and it would be better to have the height wrap the maximum height when the dropdown is open, I can quickly make that change.

Linear task: https://linear.app/comm/issue/ENG-6315/change-role-modal-on-web-changes-height-and-position-when-dropdown-is

Depends on D11064

Test Plan:
Please see the demo video below

Before:
{F1131472}

after:
{F1153405}

Reviewers: atul, inka, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10994