[native][web][lib] Extract common invite link logic
Summary: This logic was similar between platforms, and it makes sense to keep it in a single place. It becomes even more important now, because invite link blobs handling will complicate it by a lot, and we should avoid maintaining it in two places.

Depends on D10338

Test Plan: Check if accepting invite link works on web and iOS.

Reviewers: michal, bartek, rohan, inka