[lib][web][native] Add configurable authoritativeKeyserverID
Summary:
issue: [[ https://linear.app/comm/issue/ENG-6634/create-a-config-in-lib-with-the-authoritative-keyserver-info | ENG-6634 ]]
We need the constant that is currently called `ashoatKeyserverID` to be configurable. We have two different ways of configuring things on web vs native. On native we can just import .json files. On we we cannot do that (this is complicated why, see [[ https://linear.app/comm/issue/ENG-1181/refactor-libfacts-files-from-json-to-js | ENG-1181]] for more info), so we use a workaround. I followed D7371 for that

Test Plan:
Created authoritativeKeyserverID.json
```
{
  "authoritativeKeyserverID": "256"
}
```
in keyserver/facts and native/facts. Added a console log on both web and native and checked that value from the config is logged (it was set to something else than `ashoatKeyserverID`)
Reviewers:
tomek, michal
Subscribers: