[native] fix change role modal border radius style
Summary:
When a user went into the change role modal in light mode the bottom corners had a border radius that showed a darker background. To fix this I added the `cardStyleInterpolator` option to the `changeRolesScreenOptions` so that it is just a full screen modal w/o any border radius. (This is also the same convention we use in the create community full screen modal)

{F1132578}

Linear task: https://linear.app/comm/issue/ENG-6354/lower-rounded-corners-in-change-role-screen-on-mobile-need-to-be

Depends on D11002

Test Plan:
Please see demo video below

before:
{F1132577}

after:
{F1132579}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10999