[lib] Replace ashoatKeyserverID with authoritativeKeyserverID
Summary:
issue: https://linear.app/comm/issue/ENG-6635/have-ashoatkeyserverid-fetch-is-value-from-a-config
Replacing `ashoatKeyserverID` with `authoritativeKeyserverID`, which is fetched from config, based on a platform.
Because keyserver depends on web, `authoritative-keyserver.js` is run on the keyserver. On the keyserver config is not set, so we have to work aroud that. For now the keyserver shouldn't ever use `authoritativeKeyserverID`. It could in the future, but it's outside of the scope of this task.
jest tests also need `authoritativeKeyserverID`, and they also don't register configs. We use a default value for them.

Test Plan:
Tested that web and native can be run and no erros show up, tested that no errros show up on keyserver console.
Run jest tests and teset that they work and pass
Tested that it is possible to send messages, recieve them, create entries, fetch entries, log out, log in, and no errors show up

Reviewers:
michal, tomek
Subscribers: