[lib] Update genesis.id to use authoritativeKeyserverID
Summary:
issue: https://linear.app/comm/issue/ENG-6727/use-authoritativekeyserverid-in-genesis
`GENESIS` is the community holding DMs and thick threads for now. We want to have one `GENESIS` in our setups, hosted by the authoritative keyserver

Test Plan:
Tested on web and native that it is possible to create a thick thread, and no errors show up.
Checked on web that the created chat has `parentThreadID: "256|1"`, where `256` is the value of `authoritativeKeyserverID` in my setup.
That checks that `gnesis.id` is set properly, because of code in lib/actions/thread-actions.js `newTread` action, which if not provided a `parentThreadID` uses `genesis.id` (I made sure that was the case)
Reviewers:
michal, tomek
Subscribers: