[web] fix reaction tooltip not showing the correct number of users
Summary: When a message has more than 5 reactions, the reaction tooltip should only be showing the first 5 usernames then a "See more" text. However, this was not the case and we were showing all the usernames which resulted in the following visual bug.

Test Plan: Please see the screenshots below

Reviewers:

Subscribers: