[web] fix reaction tooltip not showing the correct number of users
Summary:
When a message has more than 5 reactions, the reaction tooltip should only be showing the first 5 usernames then a "See more" text. However, this was not the case and we were showing all the usernames which resulted in the following visual bug.

{F1136231}

This diff fixes this issue

Linear task: https://linear.app/comm/issue/ENG-6328/show-more-in-reaction-tooltip-is-not-working-as-expected

Test Plan:
Please see the screenshots below

before:
{F1136237}

after:
{F1136235}

confirmed tooltip still looks good with < 5 reactions

{F1136234}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11014