[keyserver] Replace ashoatKeyserverID with thisKeyserverID()
Summary:
issue: [[ https://linear.app/comm/issue/ENG-6633/replace-ashoatkeyserverid-with-calls-to-fetchidentityinfo-in-keyserver | ENG-6633 ]]
Replacing ashoatKeyserverID  with calls to fetchIdentityInfo - because this is async I had to make a couple of functions that need it be async as well.

This code has one problem - `fetchedThisKeyserverID` should be set on the fist call to `thisKeyserverID` and then this value should be returned. But that's not what happes, for some reason whenever `validateInput` is called, `thisKeyserverID` returns undefined (unless it was called very shortly after the previous call). This doesn't happen for `validateOutput`. I am not sure what is causing this. In the future it won't be so bad, because `fetchIdentityInfo` will return a value. For now this results in `registerOrLogIn` being called a lot.

Test Plan:
Tested that state sync still works by logging `infosHash` in `checkState` and compoaring with values returned by the previous code, on a client in with the same state.
Tested that the id is fetched correctly in `getInitialReduxStateResponder` and `validateOutput` (by checking that ids are prefixed correctly)

Reviewers: michal, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11050