[native] don't navigate to ExistingEthereumAccount if usingCommServicesAccessToken
Summary:
we don't have an unauth API on the identity service to check if a user already exists. maybe we can consider moving the FindUserID RPC to the unauth proto, but for now, we should just proceed to log in the existing user.

Depends on D11005

Test Plan: confirmed that we don't navigate to ExistingEthereumAccount when usingCommServicesAccessToken is enabled

Reviewers: ashoat, inka, tomek

Differential Revision: https://phab.comm.dev/D11016