[blob] Block reserved and offensive links from being created
Summary:
Copy a list of reserved usernames from keyserver - the same list is used when handling reserved links on keyserver. The same approach of copying this list is used in Identity service.

Copy a list of offensive words that is used in bad-words library. The list is modified from being an object to a plain list. The license (which is also included) allows us to do it.

Use the two lists to filter out the links that we don't want to be created.

Depends on D10992

Test Plan: Run the service locally. Modify keyserver code to not check if a link is reserved or offensive.
Check if:
1. Creating a reserved link is blocked
2. Creating an offensive link is blocked
3. Creating a link that isn't contain in these lists succeeds

Reviewers: michal, bartek, inka, varun