[web] [lib] Modify useSearchUsers to include identity search and utilize identity search on web and native
Summary:
This diff modifies the pre-existing useSearchUsers hook to conditionally call identity search if available. If not, it fallback to keyserver search. The useSearchUsers function is already used by native and web and so the identity search context provider is added.

Depends on D11000

Test Plan: flow and tested both callsites on local dev environment.

Reviewers: varun, ashoat, marcin

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11058