[web] use AddUsersList in SubchannelMembers
Summary:
This diff updates the `SubchannelMembers` to use the new `AddUsersList` that the add friends + add members modals uses. When I first initially attempted to implement this, I was running into an issue where the users in this list were not a part of the parent channel and instead I was seeing other users not apart of the parent channel or community

{F1151022}

This was caused because I was initially using `useAddMembersListUserInfos` which has different logic for selecting members than what I needed for when you create a potential sub channel. To fix this I created a new hook called `useSubchannelAddMembersListUserInfos` which uses the logic previously found in `SubchannelMembers` for selecting the correct potential users and transformed that data to what was needed for the `AddUsersList` component

Linear task: https://linear.app/comm/issue/ENG-6735/add-member-list-for-subchannel-creation-is-not-reflecting-the-correct

Depends on D10994

Test Plan:
Please see the demo video below

{F1151025}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11065