[lib][web] Bring back eslint-disable-next-line for keyserver relative import in Webpack config
Summary:
These `eslint-disable-next-line` declarations were removed in D10542, which was a diff that was automatically generated by running ESLint after adding `--report-unused-disable-directives`.

Not sure why, but the ESLint errors are showing up again for me. It seems like these errors are "unstable" (sometimes fire and sometimes not)... if that's the case, I think it's better to leave the ignore declarations in, than to take them out.

Test Plan: `yarn eslint:fix` at the repo root

Reviewers: atul

Subscribers: