[services] Move reserved_usernames to `shared`
Summary:
Move the file to `shared` and import it in two places.

This solution still requires the file to be duplicated. Solving this is more involved, as it would require storing this file in `json` format and importing it in lib, which currently isn't supported. Even if we use `fs` instead, we would still need to change the logic to be async.

Depends on D11015

Test Plan: Run both Identity and Blob services on local and Identity on Docker and check if in each case the file is accessible by console logging its content.

Reviewers: michal, bartek, varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11051