[lib] Deprecate recoveryActionSources.sqliteOpFailure
Summary:
Noticed in the last diff that this was unused. It was first introduced in D2154, and then removed by me in D5672. I neglected to deprecate the constant in that diff, unfortunately.

Following the practice elsewhere in the codebase (eg. `thread-types-enum.js`), I'm leaving the constant in but commenting it out. This lets us avoid causing any confusion by accidentally reintroducing the constant again in the future.

I could alternatively just delete it. Technically `LogInActionSource`s are passed to the keyserver, but as far as I know nothing is done with them, so there isn't too much danger in deleting them. Open to this if reviewers prefer.

Depends on D11210

Test Plan: Flow

Reviewers: inka, tomek

Subscribers: