[identity] move findUserID to unauth proto
Summary:
resolves [ENG-6782](https://linear.app/comm/issue/ENG-6782/move-finduserid-rpc-from-identity-authproto-to-identity-uanauthproto) and its subtasks

We need to be able to call this RPC from a logged out client to check if an account already exists for a given ETH wallet. In order to call it from a logged out client, it must be in the unauth service.

Test Plan: commtest test passed, flow passed

Reviewers: ashoat, bartek

Reviewed By: ashoat, bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11052