[lib] Introduce `roleIsDefaultRole`
Summary:
> Why are we introducing such a straightforward utility function in its own diff without any usage?

This is mostly because of how I've rebased and resequenced the stack. I was using the `roleIsDefaultRole` introduced by Rohan in D10116 as part of the `native` migration, but then resequenced the migration before D10116 with some modifications.

Test Plan:
Gets consumed in migration diff, careful reading.

Reviewers:
tomek, ginsu