[lib] Factor out CANCELLED_ERROR into keyserver-conn-types.js
Summary:
In a following diff, I want to use this constant from `lib/keyserver-conn/recovery-utils.js`. Since that file already imports from `keyserver-connection-handler.js`, this has the risk of introducing an import cycle. To preempt that I decided to extract the constant to a types file.

I also snuck in an unrelated minor change: I replaced the `export` at the end of `keyserver-conn-types.js` with individual `export` lines at the declarations, to make the style consistent throughout the file.

Depends on D11220

Test Plan: Flow

Reviewers: inka, tomek

Differential Revision: https://phab.comm.dev/D11221