[lib] Introduce `patchRawThreadInfosWithSpecialRole()`
Summary:
This will be used as part of `native` migration to add `specialRole` field to `RoleInfo`s. This function is `(RawThreadInfos) => RawThreadInfos` which we can pass to `updateClientDBThreadStoreThreadInfos` as the `migrationFunc`.

Test Plan:
Unit tests. Want to test this behavior outside of the running app so it was easier to ensure correct behavior w/ debugger:

Reviewers:
ginsu, tomek