[keyserver] Create invite link blobs in a migration
Summary:
In D10206 we're running this in cron, but we also should do that in a migration - this isn't strictly necessary because cron would update it later, but it is better to do that immediately.

Additionally, the function was modified so that the migration / job doesn't fail in environments where we aren't authenticated to Identity service.

Depends on D11051

Test Plan:
Run the keyserver and check if the migration creates holders in `invite_links` table.
Kill the keyserver, delete auth metadata, decrement db version, delete the holder and update invite link secret. Then run the keyserver - the migration should succeed by printing an error message - the holder shouldn't be created.

Reviewers: michal, bartek, inka

Reviewed By: inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11076