[native] persist OLM account on SQLiteQueryExecutor::storeOlmPersistData
Summary: It looks like when we refactored storeOlmPersistData as part of removing sqlite_orm, we accidentally stopped persisting the OLM account. As a result, CommCoreModule::initializeCryptoAccount() was creating a new OLM account each time it was called. This one line change fixed things.

Test Plan: called initializeCryptoAccount() multiple times and confirmed it was idempotent (returned the same identity keys) after this fix

Reviewers: ashoat, marcin, kamil

Reviewed By: marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11086