[web] fix button in add friends modal not visible on small screens
Summary:
**PLEASE NOTE THAT THIS DIFF AND SUBSEQUENT DIFFS IN THIS STACK WILL NOT BE LANDED UNTIL MORE OF THE REDESIGN IS READY SINCE THIS WILL CAUSE REGRESSIONS IN PROD**

The cta button in the add friend/block friend modal were not visible on smaller screens. This is because we were setting a fixed height of that would extend some screens. This diff fixes that bug + some copy that was requested in the linear task.

Linear task: https://linear.app/comm/issue/ENG-6387/buttons-in-add-friends-modal-not-visible-on-small-screens

Depends on D10823

Test Plan:
Please see the screenshots below

before:
{F1091666}

after:
{F1091667}

Reviewers: atul, inka, ashoat

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10849